chore: rename full_backup into full_dump and put all dump commands in dump.rs
