+++ replibyte/src/commands/mod.rs
@@ -1,3 +1,2 @@
-pub mod backup;
-pub mod restore;
+pub mod dump;
 pub mod transformer;
